+++ endtoend/components/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     testonly = True,
     srcs = [
         "beacon_node.go",
+        "boot_node.go",
         "eth1.go",
         "log.go",
         "slasher.go",
@@ -29,6 +30,5 @@ go_library(
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@io_bazel_rules_go//go/tools/bazel:go_default_library",
-        "@org_golang_x_sync//errgroup:go_default_library",
     ],
 )
